# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TravelBuddi API",
        "description": "A product of Appetiser",
        "contact": {
            "name": "Iazell",
            "email": "iazell.raper@appetiser.com.au"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{ip or domain}/api/v1",
            "description": "Base Endpoint"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user. Authenticate credentials and returns a token.",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string",
                                        "example": "testuser@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Credentials"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Login Success"
                    }
                },
                "security": []
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user. Creates new user, sends 'thank you' email and returns a token.",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string",
                                        "example": "testuser@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "Jane"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "mobile_number": {
                                        "description": "Mobile Number",
                                        "type": "string",
                                        "example": "+61 412345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "201": {
                        "description": "Request Created"
                    }
                },
                "security": []
            }
        },
        "/auth/email/check": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check email if exists. Used as a first step in user registration",
                "description": "",
                "operationId": "",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get authenticated user's information",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::getAuthUser",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Authenticated User"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/auth/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify a user after registration ",
                "description": "A 6 character verification code will be sent through email",
                "operationId": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "uid": {
                                        "description": "<b>Required</b> User ID",
                                        "type": "string",
                                        "example": "141203"
                                    },
                                    "verification_code": {
                                        "description": "<b>Required</b> Verification Code",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/auth/verification/resend": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend verification code",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::resendVerificationCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "400": {
                        "description": "Invalid Token / Bad Request"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::logout",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Logout Success"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}